BUG-5876 Enhance BGP Speaker and Peer Stats I
- Refactor existing yang model
- Expose total size (number of routes) for Adj-RIB-In and Effective-RIB-In,
  Adj-RIB-Out and a size per AFI/SAFI table for each RIB.

+++ bgp/openconfig-impl/src/main/java/org/opendaylight/protocol/bgp/openconfig/impl/moduleconfig/BGPPeerProvider.java
@@ -222,7 +222,7 @@ final class BGPPeerProvider {
         bgpPeerBuilder.setHost(ipAdress);
         final Timers timers = neighbor.getTimers();
         if (timers != null && timers.getConfig() != null && timers.getConfig().getHoldTime() != null) {
-            bgpPeerBuilder.setHoldtimer(neighbor.getTimers().getConfig().getHoldTime().shortValue());
+            bgpPeerBuilder.setHoldtimer(neighbor.getTimers().getConfig().getHoldTime().intValue());
         }
         final Transport transport = neighbor.getTransport();
         if (transport != null && transport.getConfig() != null && transport.getConfig().isPassiveMode() != null) {
